+++ Changelog.md
@@ -10,6 +10,7 @@
 - Added Laravel 12 Support
 - Added the `gtin` field to the product and master product variant forms
 - Added sorting and filtering options to the Product index
+- Added ETA fields to the Shipping Method forms
 
 ## 4.2.0
 ##### 2024-12-15